package parameter validation code
