+++ plugins/nf-schema/src/main/nextflow/validation/ValidationExtension.groovy
@@ -35,6 +35,7 @@ import nextflow.validation.exceptions.SchemaValidationException
 import nextflow.validation.help.HelpMessage
 import nextflow.validation.validators.JsonSchemaValidator
 import nextflow.validation.samplesheet.SamplesheetConverter
+import nextflow.validation.parameters.ParameterValidator
 import static nextflow.validation.utils.Colors.logColors
 import static nextflow.validation.utils.Files.paramsLoad
 import static nextflow.validation.utils.Common.getSchemaPath
@@ -51,85 +52,6 @@ import static nextflow.validation.utils.Common.findDeep
 @CompileStatic
 class ValidationExtension extends PluginExtensionPoint {
 
-    final List<String> NF_OPTIONS = [
-            // Options for base `nextflow` command
-            'bg',
-            'c',
-            'C',
-            'config',
-            'd',
-            'D',
-            'dockerize',
-            'h',
-            'log',
-            'q',
-            'quiet',
-            'syslog',
-            'v',
-
-            // Options for `nextflow run` command
-            'ansi',
-            'ansi-log',
-            'bg',
-            'bucket-dir',
-            'c',
-            'cache',
-            'config',
-            'dsl2',
-            'dump-channels',
-            'dump-hashes',
-            'E',
-            'entry',
-            'latest',
-            'lib',
-            'main-script',
-            'N',
-            'name',
-            'offline',
-            'params-file',
-            'pi',
-            'plugins',
-            'poll-interval',
-            'pool-size',
-            'profile',
-            'ps',
-            'qs',
-            'queue-size',
-            'r',
-            'resume',
-            'revision',
-            'stdin',
-            'stub',
-            'stub-run',
-            'test',
-            'w',
-            'with-charliecloud',
-            'with-conda',
-            'with-dag',
-            'with-docker',
-            'with-mpi',
-            'with-notification',
-            'with-podman',
-            'with-report',
-            'with-singularity',
-            'with-timeline',
-            'with-tower',
-            'with-trace',
-            'with-weblog',
-            'without-docker',
-            'without-podman',
-            'work-dir'
-    ]
-
-    private List<String> errors = []
-    private List<String> warnings = []
-
-    // The amount of parameters hidden (for help messages)
-    private Integer hiddenParametersCount = 0
-
-    // The length of the terminal
-    private Integer terminalLength = System.getenv("COLUMNS")?.toInteger() ?: 100
-
     // The configuration class
     private ValidationConfig config
 
@@ -139,6 +61,7 @@ class ValidationExtension extends PluginExtensionPoint {
     @Override
     protected void init(Session session) {
         def plugins = session?.config?.navigate("plugins") as ArrayList
+        // TODO move the warning to an observer
         if(plugins?.contains("nf-schema")) {
             log.warn("""
 !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!
@@ -164,20 +87,6 @@ class ValidationExtension extends PluginExtensionPoint {
         def Map params = (Map)session.params ?: [:]
         config = new ValidationConfig(session?.config?.navigate('validation') as Map, params)
 
-    }
-
-    boolean hasErrors() { errors.size()>0 }
-    List<String> getErrors() { errors }
-
-    boolean hasWarnings() { warnings.size()>0 }
-    List<String> getWarnings() { warnings }
-
-    //
-    // Find a value in a nested map
-    //
-    def findDeep(Map m, String key) {
-        if (m.containsKey(key)) return m[key]
-        m.findResult { k, v -> v instanceof Map ? findDeep(v, key) : null }
     }
 
     @Function
@@ -225,18 +134,17 @@ class ValidationExtension extends PluginExtensionPoint {
     //
     // Initialise expected params if not present
     //
-    Map initialiseExpectedParams(Map params) {
+    private Map initialiseExpectedParams(Map params) {
         addExpectedParams().each { param ->
             params[param] = false
         }
         return params
     }
 
-
     //
     // Add expected params
     //
-    List addExpectedParams() {
+    private List addExpectedParams() {
         def List expectedParams = [
             config.help.shortParameter,
             config.help.fullParameter,
@@ -254,106 +162,14 @@ class ValidationExtension extends PluginExtensionPoint {
     void validateParameters(
         Map options = null
     ) {
-
         def Map params = initialiseExpectedParams(session.params)
-        def String baseDir = session.baseDir.toString()
-        def String schemaFilename = options?.containsKey('parameters_schema') ? options.parameters_schema as String : config.parametersSchema
-        log.debug "Starting parameters validation"
-
-        // Clean the parameters
-        def cleanedParams = cleanParameters(params)
-        // Convert to JSONObject
-        def paramsJSON = new JSONObject(new JsonBuilder(cleanedParams).toString())
-
-        //=====================================================================//
-        // Check for nextflow core params and unexpected params
-        def slurper = new JsonSlurper()
-        def Map parsed = (Map) slurper.parse( Path.of(getSchemaPath(baseDir, schemaFilename)) )
-        // $defs is the adviced keyword for definitions. Keeping defs in for backwards compatibility
-        def Map schemaParams = (Map) (parsed.get('$defs') ?: parsed.get("defs"))
-        def specifiedParamKeys = params.keySet()
-
-        // Collect expected parameters from the schema
-        def enumsTuple = collectEnums(schemaParams)
-        def List expectedParams = (List) enumsTuple[0] + addExpectedParams()
-        def Map enums = (Map) enumsTuple[1]
-        // Collect expected parameters from the schema when parameters are specified outside of "$defs"
-        if (parsed.containsKey('properties')) {
-            def enumsTupleTopLevel = collectEnums(['top_level': ['properties': parsed.get('properties')]])
-            expectedParams += (List) enumsTupleTopLevel[0]
-            enums += (Map) enumsTupleTopLevel[1]
-        }
-
-        //=====================================================================//
-        for (String specifiedParam in specifiedParamKeys) {
-            // nextflow params
-            if (NF_OPTIONS.contains(specifiedParam)) {
-                errors << "You used a core Nextflow option with two hyphens: '--${specifiedParam}'. Please resubmit with '-${specifiedParam}'".toString()
-            }
-            // unexpected params
-            def expectedParamsLowerCase = expectedParams.collect{ it -> 
-                def String p = it
-                p.replace("-", "").toLowerCase() 
-            }
-            def specifiedParamLowerCase = specifiedParam.replace("-", "").toLowerCase()
-            def isCamelCaseBug = (specifiedParam.contains("-") && !expectedParams.contains(specifiedParam) && expectedParamsLowerCase.contains(specifiedParamLowerCase))
-            if (!expectedParams.contains(specifiedParam) && !config.ignoreParams.contains(specifiedParam) && !isCamelCaseBug) {
-                if (config.failUnrecognisedParams) {
-                    errors << "* --${specifiedParam}: ${params[specifiedParam]}".toString()
-                } else {
-                    warnings << "* --${specifiedParam}: ${params[specifiedParam]}".toString()
-                }
-            }
-        }
-
-        //=====================================================================//
-        // Validate parameters against the schema
-        def String schema_string = Files.readString( Path.of(getSchemaPath(baseDir, schemaFilename)) )
-        def validator = new JsonSchemaValidator(config)
-
-        // check for warnings
-        if( this.hasWarnings() ) {
-            def msg = "The following invalid input values have been detected:\n\n" + this.getWarnings().join('\n').trim() + "\n\n"
-            log.warn(msg)
-        }
-
-        // Colors
-        def colors = logColors(config.monochromeLogs)
-
-        // Validate
-        List<String> validationErrors = validator.validate(paramsJSON, schema_string)
-        this.errors.addAll(validationErrors)
-        def List<String> modifiedIgnoreParams = config.ignoreParams.collect { param -> "* --${param}" as String }
-        def List<String> filteredErrors = errors.findAll { error -> 
-            return modifiedIgnoreParams.find { param -> error.startsWith(param) } == null
-        }
-        if (filteredErrors.size() > 0) {
-            def msg = "${colors.red}The following invalid input values have been detected:\n\n" + filteredErrors.join('\n').trim() + "\n${colors.reset}\n"
-            log.error("Validation of pipeline parameters failed!")
-            throw new SchemaValidationException(msg, this.getErrors())
-        }
-
-        log.debug "Finishing parameters validation"
-    }
-
-    //
-    // Function to collect enums (options) of a parameter and expected parameters (present in the schema)
-    //
-    Tuple collectEnums(Map schemaParams) {
-        def expectedParams = []
-        def enums = [:]
-        for (group in schemaParams) {
-            def Map properties = (Map) group.value['properties']
-            for (p in properties) {
-                def String key = (String) p.key
-                expectedParams.push(key)
-                def Map property = properties[key] as Map
-                if (property.containsKey('enum')) {
-                    enums[key] = property['enum']
-                }
-            }
-        }
-        return new Tuple (expectedParams, enums)
+        def validator = new ParameterValidator(config)
+        validator.validateParametersMap(
+            options,
+            params,
+            session.baseDir.toString(),
+            addExpectedParams()
+        )
     }
 
     //
@@ -548,35 +364,5 @@ Please contact the pipeline maintainer(s) if you see this warning as a user.
             }
         }
         return returnMap
-    }
-
-    //
-    // Clean and check parameters relative to Nextflow native classes
-    //
-    private Map cleanParameters(Map params) {
-        def Map new_params = (Map) params.getClass().newInstance(params)
-        for (p in params) {
-            // remove anything evaluating to false
-            if (!p['value'] && p['value'] != 0) {
-                new_params.remove(p.key)
-            }
-            // Cast MemoryUnit to String
-            if (p['value'] instanceof MemoryUnit) {
-                new_params.replace(p.key, p['value'].toString())
-            }
-            // Cast Duration to String
-            if (p['value'] instanceof Duration) {
-                new_params.replace(p.key, p['value'].toString())
-            }
-            // Cast LinkedHashMap to String
-            if (p['value'] instanceof LinkedHashMap) {
-                new_params.replace(p.key, p['value'].toString())
-            }
-            // Parsed nested parameters
-            if (p['value'] instanceof Map) {
-                new_params.replace(p.key, cleanParameters(p['value'] as Map))
-            }
-        }
-        return new_params
     }
 }
